+++ src/assets/icons/filter/filterIcon.tsx
@@ -1,0 +1,9 @@
+export const FilterIcon = () => {
+  return (
+    <div>
+      <svg width="18" height="12" viewBox="0 0 18 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M7 12V10H11V12H7ZM3 7V5H15V7H3ZM0 2V0H18V2H0Z" fill="#191A1C"/>
+      </svg>
+    </div>
+  );
+};
